First commit
In this commit we have the main game logic that has the player and the obstacles along with detecting when a player collides with the obstacles. In addition to that, we have the background other assets in the game. Moving forward just need to add a sprite for the obstacles, fix the scale of the speed of obstacles, match the speed of obstacles with background scroll, gameover logic with menus, and a few other improvements like choosing characters.